# openapi.json
{
  "components": {
    "schemas": {
      "Body_convert_convert_post": {
        "properties": {
          "biochef_workflow": {
            "title": "Biochef Workflow",
            "type": "string"
          },
          "files": {
            "items": {
              "contentMediaType": "application/octet-stream",
              "type": "string"
            },
            "title": "Files",
            "type": "array"
          }
        },
        "required": [
          "biochef_workflow",
          "files"
        ],
        "title": "Body_convert_convert_post",
        "type": "object"
      },
      "Body_submit_run_runs_post": {
        "properties": {
          "biochef_workflow": {
            "title": "Biochef Workflow",
            "type": "string"
          },
          "files": {
            "items": {
              "contentMediaType": "application/octet-stream",
              "type": "string"
            },
            "title": "Files",
            "type": "array"
          }
        },
        "required": [
          "biochef_workflow",
          "files"
        ],
        "title": "Body_submit_run_runs_post",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/convert": {
      "post": {
        "description": "The editor's synchronous contract, sharing the execution limit.\n\nKept as it was because it is the contract the editor speaks today. /runs is\nthe same work without the wait.",
        "operationId": "convert_convert_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_convert_convert_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Convert"
      }
    },
    "/runs": {
      "post": {
        "description": "Accept a workflow and answer immediately with something to poll.\n\nThe uploads are read here, while the request is still open. By the time the\nwork runs there is no request left to read them from -- which is the whole\ndifference between this and /convert, and the reason it cannot simply call\nthe same handler in the background.",
        "operationId": "submit_run_runs_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_submit_run_runs_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "Agent capacity reached; retry later"
          }
        },
        "summary": "Submit Run"
      }
    },
    "/runs/{run_id}": {
      "get": {
        "description": "Where a run has got to, and its outputs once it is COMPLETE.",
        "operationId": "get_run_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run"
      }
    },
    "/runs/{run_id}/cancel": {
      "post": {
        "description": "Ask a run to stop, and end the processes doing it.\n\nThe path is WES's, so exposing this as a WES endpoint later (F5) does not\nmove it.\n\nTwo shapes of run, and they differ. One waiting for a slot has executed\nnothing, so cancelling it is a matter of state: it settles CANCELED when its\nturn comes and it declines to start. One that is running has a process\ngroup, and that group is ended -- the tool and children that remain in the\ngroup, exactly as the timeout does it.\n\nThe reply is normally CANCELING while the worker tidies up. It may already\nbe CANCELED if the worker finishes before the response; CANCELED is only\nrecorded after cleanup.",
        "operationId": "cancel_run_runs__run_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Run"
      }
    },
    "/runs/{run_id}/logs": {
      "get": {
        "description": "Run-level output and separately captured output from each executed node.\n\n`stdout` and `stderr` are Snakemake's run-wide streams. `node_logs` has\nstdout and stderr from separate files for each rule that started, including\na failing tool's actual stderr. Stdout already directed into a declared\nscientific output remains in that output and is not copied into a log.\n`failed_steps` is a diagnostic index of Snakemake-style error headings in the\nrun-wide stderr; tools can imitate those headings, so it is not proof of\nfailure or a substitute for `node_logs`.\n\nLogs are recorded after the workflow process exits, before output collection\nand the run's terminal state. They are not available during execution.\nEach stream's retained per-node tails share BIOCHEF_MAX_LOG_BYTES across\nnodes; log files on disk can grow until the run ends.",
        "operationId": "get_run_logs_runs__run_id__logs_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run Logs"
      }
    }
  }
}